Begin offloading to helpers.py
