+++ config/config.json
@@ -7,5 +7,5 @@
     "log_file": "ispyra.log",
     "log_messages": true,
     "log_commands": true,
-    "version": "2.0.0"
+    "version": "2.1.0"
 }
